+++ packages/cloud-cognitive/src/components/DataSpreadsheet/DataSpreadsheet.js
@@ -51,6 +51,7 @@ import { selectAllCells } from './utils/selectAllCells';
 import { handleEditSubmit } from './utils/handleEditSubmit';
 import { handleActiveCellInSelectionEnter } from './utils/handleActiveCellInSelectionEnter';
 import { handleActiveCellInSelectionTab } from './utils/handleActiveCellInSelectionTab';
+import { handleCellDeletion } from './utils/handleCellDeletion';
 // cspell:words rowcount colcount
 
 // The block part of our conventional BEM class names (blockClass__E--M).
@@ -66,6 +67,7 @@ const defaults = {
   onDataUpdate: Object.freeze(() => {}),
   onActiveCellChange: Object.freeze(() => {}),
   onSelectionAreaChange: Object.freeze(() => {}),
+  theme: 'light',
 };
 
 /**
@@ -84,6 +86,9 @@ export let DataSpreadsheet = React.forwardRef(
       id,
       onActiveCellChange = defaults.onActiveCellChange,
       onSelectionAreaChange = defaults.onSelectionAreaChange,
+      selectAllAriaLabel,
+      spreadsheetAriaLabel,
+      theme,
       totalVisibleColumns,
 
       // Collect any other property values passed in.
@@ -151,13 +156,13 @@ export let DataSpreadsheet = React.forwardRef(
 
     // Update the spreadsheet data after editing a cell
     const updateData = useCallback(
-      (rowIndex, columnId) => {
+      (rowIndex, columnId, newValue) => {
         onDataUpdate((prev) =>
           prev.map((row, index) => {
             if (index === rowIndex) {
               return {
                 ...prev[rowIndex],
-                [columnId]: cellEditorValue,
+                [columnId]: cellEditorValue || newValue,
               };
             }
             return row;
@@ -412,6 +417,13 @@ export let DataSpreadsheet = React.forwardRef(
             usingMac,
           });
         }
+        const deleteParams = {
+          selectionAreas,
+          currentMatcher,
+          rows,
+          setActiveCellContent,
+          updateData,
+        };
         // Allow arrow key navigation if there are less than two activeKeys OR
         // if one of the activeCellCoordinates is in a header position
         if (
@@ -420,6 +432,16 @@ export let DataSpreadsheet = React.forwardRef(
           activeCellCoordinates.column === 'header'
         ) {
           switch (key) {
+            // Backspace
+            case 'Backspace': {
+              handleCellDeletion(deleteParams);
+              break;
+            }
+            // Delete
+            case 'Delete': {
+              handleCellDeletion(deleteParams);
+              break;
+            }
             // Enter
             case 'Enter': {
               handleActiveCellInSelectionEnter({
@@ -577,6 +599,7 @@ export let DataSpreadsheet = React.forwardRef(
         handleHomeEndKey,
         keysPressedList,
         usingMac,
+        updateData,
       ]
     );
 
@@ -628,6 +651,11 @@ export let DataSpreadsheet = React.forwardRef(
         handleRowColumnHeaderClick({ isKeyboard: false, index: indexValue });
       }
       return;
+    };
+
+    // Mouse up on active cell
+    const handleActiveCellMouseUp = () => {
+      setClickAndHoldActive(false);
     };
 
     // Mouse down on active cell
@@ -794,6 +822,7 @@ export let DataSpreadsheet = React.forwardRef(
           `${blockClass}--interactive-cell-element`,
           {
             [`${blockClass}__container-has-focus`]: containerHasFocus,
+            [`${blockClass}__${theme}`]: theme === 'dark',
           }
         )}
         ref={spreadsheetRef}
@@ -801,6 +830,7 @@ export let DataSpreadsheet = React.forwardRef(
         tabIndex={0}
         aria-rowcount={rows?.length || 0}
         aria-colcount={columns?.length || 0}
+        aria-label={spreadsheetAriaLabel}
         onKeyDown={handleKeyPress}
         onFocus={() => setContainerHasFocus(true)}
       >
@@ -826,6 +856,7 @@ export let DataSpreadsheet = React.forwardRef(
             setHeaderCellHoldActive={setHeaderCellHoldActive}
             headerCellHoldActive={headerCellHoldActive}
             visibleColumns={visibleColumns}
+            selectAllAriaLabel={selectAllAriaLabel}
           />
 
           {/* BODY */}
@@ -864,6 +895,7 @@ export let DataSpreadsheet = React.forwardRef(
           />
           <button
             onMouseDown={handleActiveCellMouseDown}
+            onMouseUp={handleActiveCellMouseUp}
             onClick={handleActiveCellClick}
             onKeyDown={handleActiveCellKeyDown}
             onDoubleClick={handleActiveCellDoubleClick}
@@ -904,7 +936,7 @@ export let DataSpreadsheet = React.forwardRef(
             labelText=""
             aria-labelledby={
               activeCellCoordinates
-                ? `#${blockClass}__cell--${activeCellCoordinates?.row}--${activeCellCoordinates?.column}`
+                ? `${blockClass}__cell--${activeCellCoordinates?.row}--${activeCellCoordinates?.column}`
                 : null
             }
             className={cx(
@@ -988,6 +1020,21 @@ DataSpreadsheet.propTypes = {
    * The event handler that is called when the selection area values change
    */
   onSelectionAreaChange: PropTypes.func,
+
+  /**
+   * The aria label applied to the Select all button
+   */
+  selectAllAriaLabel: PropTypes.string.isRequired,
+
+  /**
+   * The aria label applied to the Data spreadsheet component
+   */
+  spreadsheetAriaLabel: PropTypes.string.isRequired,
+
+  /**
+   * The theme the DataSpreadsheet should use (only used to render active cell/selection area colors on dark theme)
+   */
+  theme: PropTypes.oneOf(['light', 'dark']),
 
   /**
    * The total number of columns to be initially visible, additional columns will be rendered and