+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid.stories.js
@@ -7,19 +7,20 @@
  */
 
 import React, { useState, useEffect } from 'react';
-// TODO: import action to handle events if required.
-// import { action } from '@storybook/addon-actions';
-import { useColumnOrder } from 'react-table';
 import { range, makeData, newPersonWithTwoLines } from './utils/makeData';
+import { getInlineEditColumns } from './utils/getInlineEditColumns';
 
 import { getStoryTitle } from '../../global/js/utils/story-helper';
 
+import { action } from '@storybook/addon-actions';
 import {
   Activity16,
   Restart16,
   Download16,
   Filter16,
   Add16,
+  Edit16,
+  TrashCan16,
 } from '@carbon/icons-react';
 import { DataTable, Button, Pagination } from 'carbon-components-react';
 import {
@@ -39,6 +40,8 @@ import {
   useColumnCenterAlign,
   useStickyColumn,
   useActionsColumn,
+  useColumnOrder,
+  useInlineEdit,
 } from '.';
 
 import {
@@ -54,14 +57,11 @@ import cx from 'classnames';
 
 import styles from './_storybook-styles.scss';
 import { SidePanel } from '../SidePanel';
+import { ButtonMenu, ButtonMenuItem } from '../ButtonMenu';
 
 export default {
   title: getStoryTitle(Datagrid.displayName),
   component: Datagrid,
-  // TODO: Define argTypes for props not represented by standard JS types.
-  // argTypes: {
-  //   egProp: { control: 'color' },
-  // },
   parameters: {
     styles,
     docs: {
@@ -96,7 +96,6 @@ const defaultHeader = [
   {
     Header: 'First Name',
     accessor: 'firstName',
-    sticky: 'left',
   },
   {
     Header: 'Last Name',
@@ -157,7 +156,17 @@ const defaultHeader = [
 const { TableBatchAction, TableBatchActions } = DataTable;
 
 export const BasicUsage = () => {
-  const columns = React.useMemo(() => defaultHeader, []);
+  const columns = React.useMemo(
+    () => [
+      ...defaultHeader,
+      {
+        Header: 'Someone 11',
+        accessor: 'someone11',
+        multiLineWrap: true,
+      },
+    ],
+    []
+  );
   const [data] = useState(makeData(10));
   const datagridState = useDatagrid({
     columns,
@@ -447,6 +456,21 @@ export const ClickableRow = () => {
   );
 };
 
+export const InlineEdit = () => {
+  const [data, setData] = useState(makeData(10));
+  const columns = React.useMemo(() => getInlineEditColumns(), []);
+  const datagridState = useDatagrid(
+    {
+      columns,
+      data,
+      onDataUpdate: setData,
+      DatagridActions,
+    },
+    useInlineEdit
+  );
+  return <Datagrid datagridState={{ ...datagridState }} />;
+};
+
 const DataTableSidePanelContent = (selectedRowValues) => {
   const { rowData } = selectedRowValues;
 
@@ -533,8 +557,12 @@ export const SelectableRow = () => {
     {
       columns,
       data,
+      DatagridActions,
+      batchActions: true,
+      toolbarBatchActions: getBatchActions(),
     },
-    useSelectRows
+    useSelectRows,
+    useStickyColumn
   );
 
   return <Datagrid datagridState={{ ...datagridState }} />;
@@ -556,7 +584,8 @@ export const RadioSelect = () => {
         },
       },
     },
-    useSelectRows
+    useSelectRows,
+    useStickyColumn
   );
 
   return <Datagrid datagridState={{ ...datagridState }} />;
@@ -678,6 +707,7 @@ const DatagridActions = (datagridState) => {
       bottom: '-37px',
     },
   };
+
   return (
     isNothingSelected &&
     (useDenseHeader && useDenseHeader ? (
@@ -759,6 +789,20 @@ const DatagridActions = (datagridState) => {
               <CustomizeColumnsButton />
             </div>
           )}
+          <ButtonMenu label="Primary button" renderIcon={Add16}>
+            <ButtonMenuItem
+              itemText="Option 1"
+              onClick={action(`Click on ButtonMenu Option 1`)}
+            />
+            <ButtonMenuItem
+              itemText="Option 2"
+              onClick={action(`Click on ButtonMenu Option 2`)}
+            />
+            <ButtonMenuItem
+              itemText="Option 3"
+              onClick={action(`Click on ButtonMenu Option 3`)}
+            />
+          </ButtonMenu>
         </TableToolbarContent>
       </>
     ))
@@ -774,6 +818,18 @@ export const DatagridActionsToolbar = () => {
       data,
       DatagridActions,
       DatagridBatchActions,
+      rowSizeProps: {
+        labels: {
+          rowSizeLabels: {
+            xl: 'Extra large',
+            lg: 'Large (default)',
+            md: 'Medium',
+            sm: 'Small',
+            xs: 'Extra small',
+          },
+          legendText: 'Row height',
+        },
+      },
     },
     useSelectRows
   );
@@ -811,7 +867,7 @@ export const SelectItemsInAllPages = () => {
     <>
       <Datagrid datagridState={{ ...datagridState }} />
       <h3>Doc in Notes...</h3>
-      <p>{`Are all selected across all pages? - ${areAllSelected}`}</p>
+      <p>{`Are all entries selected across all pages? - ${areAllSelected}`}</p>
     </>
   );
 };
@@ -832,6 +888,21 @@ export const CustomizingColumns = () => {
       customizeColumnsProps: {
         onSaveColumnPrefs: (newColDefs) => {
           console.log(newColDefs);
+        },
+        labels: {
+          findColumnPlaceholderLabel: 'Find column',
+          resetToDefaultLabel: 'Reset to default',
+          customizeModalHeadingLabel: 'Customize display',
+          primaryButtonTextLabel: 'Save',
+          secondaryButtonTextLabel: 'Cancel',
+          instructionsLabel:
+            'Deselect columns to hide them. Click and drag the white box to reorder the columns. These specifications will be saved and persist if you leave and return to the data table.',
+          iconTooltipLabel: 'Customize columns',
+          assistiveTextInstructionsLabel:
+            'Press space bar to toggle drag drop mode, use arrow keys to move selected elements.',
+          assistiveTextDisabledInstructionsLabel:
+            'Reordering columns are disabled because they are filtered currently.',
+          selectAllLabel: 'Column name',
         },
       },
       DatagridActions,
@@ -940,6 +1011,7 @@ const DatagridBatchActions = (datagridState) => {
   const totalSelected = selectedFlatRows && selectedFlatRows.length;
   const onBatchAction = () => alert('Batch action');
   const actionName = 'Action';
+
   return (
     <TableBatchActions
       shouldShowBatchActions={totalSelected > 0}
@@ -953,6 +1025,44 @@ const DatagridBatchActions = (datagridState) => {
   );
 };
 
+const getBatchActions = () => {
+  return [
+    {
+      label: 'Duplicate',
+      renderIcon: Add16,
+      onClick: action('Clicked batch action button'),
+    },
+    {
+      label: 'Add',
+      renderIcon: Add16,
+      onClick: action('Clicked batch action button'),
+    },
+    {
+      label: 'Select all',
+      renderIcon: Add16,
+      onClick: action('Clicked batch action button'),
+      type: 'select_all',
+    },
+    {
+      label: 'Publish to catalog',
+      renderIcon: Add16,
+      onClick: action('Clicked batch action button'),
+    },
+    {
+      label: 'Download',
+      renderIcon: Add16,
+      onClick: action('Clicked batch action button'),
+    },
+    {
+      label: 'Delete',
+      renderIcon: Add16,
+      onClick: action('Clicked batch action button'),
+      hasDivider: true,
+      kind: 'danger',
+    },
+  ];
+};
+
 export const BatchActions = () => {
   const columns = React.useMemo(() => defaultHeader, []);
   const [data] = useState(makeData(10));
@@ -960,10 +1070,13 @@ export const BatchActions = () => {
     {
       columns,
       data,
+      batchActions: true,
+      toolbarBatchActions: getBatchActions(),
       DatagridActions,
       DatagridBatchActions,
     },
-    useSelectRows
+    useSelectRows,
+    useSelectAllWithToggle
   );
 
   return <Datagrid datagridState={{ ...datagridState }} />;
@@ -987,8 +1100,6 @@ export const DisableSelectRow = () => {
 
   return <Datagrid datagridState={{ ...datagridState }} />;
 };
-
-// export { StickyActionsColumn };
 
 NestedRows.story = {
   parameters: {
@@ -1096,6 +1207,10 @@ export const StickyActionsColumn = () => {
     {
       columns,
       data,
+      batchActions: true,
+      toolbarBatchActions: getBatchActions(),
+      DatagridActions,
+      DatagridBatchActions,
       rowActions: [
         {
           id: 'edit',
@@ -1119,6 +1234,65 @@ export const StickyActionsColumn = () => {
           id: 'delete',
           itemText: 'Delete',
           hasDivider: true,
+          isDelete: true,
+          onClick: onActionClick,
+        },
+      ],
+    },
+    useStickyColumn,
+    useActionsColumn,
+    useSelectRows,
+    useSelectAllWithToggle
+  );
+  return (
+    <Wrapper>
+      <h3>{msg}</h3>
+      <Datagrid datagridState={{ ...datagridState }} />
+      <p>More details documentation check the Notes section below</p>
+    </Wrapper>
+  );
+};
+
+export const RowActionButton = () => {
+  const columns = React.useMemo(
+    () => [
+      ...defaultHeader,
+      {
+        Header: '',
+        accessor: 'actions',
+        sticky: 'right',
+        width: 90,
+        isAction: true,
+      },
+    ],
+    []
+  );
+  const [data] = useState(makeData(10));
+  const [msg, setMsg] = useState('click action menu');
+  const onActionClick = (actionId, row) => {
+    console.log(onActionClick);
+    const { original } = row;
+    setMsg(
+      `Clicked [${actionId}] on row: <${original.firstName} ${original.lastName}>`
+    );
+  };
+
+  const datagridState = useDatagrid(
+    {
+      columns,
+      data,
+      rowActions: [
+        {
+          id: 'edit',
+          itemText: 'Edit',
+          icon: Edit16,
+          onClick: onActionClick,
+        },
+
+        {
+          id: 'delete',
+          itemText: 'Delete',
+          icon: TrashCan16,
           isDelete: true,
           onClick: onActionClick,
         },