+++ packages/cloud-cognitive/src/components/Datagrid/index.js
@@ -21,3 +21,5 @@ export { default as useActionsColumn } from './useActionsColumn';
 export { default as useCustomizeColumns } from './useCustomizeColumns';
 export { default as useSelectAllWithToggle } from './useSelectAllToggle';
 export { default as useColumnCenterAlign } from './useColumnCenterAlign';
+export { default as useColumnOrder } from './useColumnOrder';
+export { default as useInlineEdit } from './useInlineEdit';