+++ packages/cloud-cognitive/src/components/Datagrid/utils/makeData.js
@@ -7,6 +7,13 @@
 
 import React from 'react';
 import namor from 'namor';
+import { inlineEditSelectItems } from './getInlineEditColumns';
+
+const getRandomInteger = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
 
 export const makeData = (...lens) => {
   const makeDataLevel = (depth = 0) => {
@@ -30,6 +37,12 @@ export const range = (len) => {
 
 const newPerson = () => {
   const statusChance = Math.random();
+  const initialChartTypeIndex = getRandomInteger(0, 2);
+  const activeSinceDate = new Date();
+  let yesterdayDate = new Date();
+  yesterdayDate.setDate(yesterdayDate.getDate() - 1);
+  let twoDaysAgoDate = new Date();
+  twoDaysAgoDate.setDate(twoDaysAgoDate.getDate() - 2);
   return {
     firstName: namor.generate({ words: 1, numbers: 0 }),
     lastName: namor.generate({ words: 1, numbers: 0 }),
@@ -46,7 +59,7 @@ const newPerson = () => {
     someone8: namor.generate({ words: 1, numbers: 0 }),
     someone9: namor.generate({ words: 1, numbers: 0 }),
     someone10: namor.generate({ words: 1, numbers: 0 }),
-    someone11: namor.generate({ words: 1, numbers: 0 }),
+    someone11: namor.generate({ words: 4, numbers: 0 }),
     someone12: namor.generate({ words: 1, numbers: 0 }),
     someone13: namor.generate({ words: 1, numbers: 0 }),
     someone14: namor.generate({ words: 1, numbers: 0 }),
@@ -62,6 +75,18 @@ const newPerson = () => {
         : statusChance > 0.33
         ? 'complicated'
         : 'single',
+    chartType:
+      initialChartTypeIndex === 0
+        ? inlineEditSelectItems[0]
+        : initialChartTypeIndex === 1
+        ? inlineEditSelectItems[1]
+        : inlineEditSelectItems[2],
+    activeSince:
+      statusChance > 0.66
+        ? activeSinceDate
+        : statusChance > 0.33
+        ? yesterdayDate
+        : twoDaysAgoDate,
   };
 };